Fixed logout, updated Dockerfile

+++ app/Controllers/Http/FriendsController.ts
@@ -7,6 +7,10 @@ Controller for Friends feature
 
 export default class FriendsController {
     public async store({ params, response, auth }: HttpContextContract) {
+        if(!auth.isLoggedIn) {
+            return response.status(401)
+        }
+
         const friend = new Friend
         friend.userId = auth.user!.id
         friend.friendId = params.user.id
@@ -16,10 +20,15 @@ export default class FriendsController {
     }
 
     public async destroy({ params, response, auth }) {
+        if(!auth.isLoggedIn) {
+            return response.status(401)
+        }
+
         const friend = Friend
             .query()
             .where('user_id', auth.user!.id)
-            .where('following_id', params.user_id);
+            .where('following_id', params.user_id)
+
         await friend.delete()
         return response.status(200)
     }
